+++ src/main/java/net/meander/subtlyd/network/syncher/SynchedEntityDataSD.java
@@ -13,6 +13,7 @@ public class SynchedEntityDataSD {
     public static final EntityDataAccessor<Float> DATA_ID_DAMAGE = SynchedEntityData.defineId(TentEntity.class, EntityDataSerializers.FLOAT);
     public static final EntityDataAccessor<Boolean> DATA_ID_ZOMBIE_LEADER = SynchedEntityData.defineId(Zombie.class, EntityDataSerializers.BOOLEAN);
     public static final EntityDataAccessor<Boolean> DATA_ID_WITHER_DIVE = SynchedEntityData.defineId(WitherBoss.class, EntityDataSerializers.BOOLEAN);
+    public static EntityDataAccessor<Boolean> DATA_ID_CHARGED_TRIDENT;
 
     public static void createEntityData() {}
 }